Created augmentation.py file
This file is to combine the query of the user with the
returned abstracts to form a context to pass a cleaner
piece of text to the generative model.

In the tests directory are tests for valid queries and
queries that did not return any results to ensure that
errors are handled graciously.